+++ DrawingApp/DrawingApp/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
