+++ DrawingApp/DrawingApp/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
