+++ DrawingApp/DrawingApp/Resources/Assets.xcassets/Contents.json
