+++ DrawingApp/DrawingApp/Resources/Storyboards/Base.lproj/LaunchScreen.storyboard
